+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - GitLab
   - Make general Merge Request template the default. (!906)
   - Add documentation redirect instructions to Release merge request template. (!906)
+- Tests - Fail pytest when using deprecated code from `quantify-scheduler` unintentionally. (!899)
 
 ## 0.18.0 (2023-12-22)